+++ tests/py/test_privacy_json.py
@@ -6,7 +6,6 @@ from gratipay.models.participant import Participant
 
 
 class Tests(Harness):
-
     def setUp(self):
         Harness.setUp(self)
         self.make_participant('alice', claimed_time='now')
@@ -20,7 +19,11 @@ class Tests(Harness):
     def test_participant_can_get_their_privacy_settings(self):
         response = self.hit_privacy('GET')
         actual = json.loads(response.body)
-        assert actual == {'is_searchable': True}
+        assert actual == {
+            'is_searchable': True,
+            'anonymous_giving': False,
+            'anonymous_receiving': False
+        }
 
     def test_participant_can_toggle_is_searchable(self):
         response = self.hit_privacy('POST', data={'toggle': 'is_searchable'})
@@ -32,6 +35,30 @@ class Tests(Harness):
         response = self.hit_privacy('POST', data={'toggle': 'is_searchable'})
         actual = json.loads(response.body)
         assert actual['is_searchable'] is True
+
+    def test_participant_can_toggle_anonymous_giving(self):
+        response = self.hit_privacy('POST', data={'toggle': 'anonymous_giving'})
+        actual = json.loads(response.body)
+        assert actual['giving'] is True
+
+    def test_participant_can_toggle_anonymous_giving_back(self):
+        response = self.hit_privacy('POST', data={'toggle': 'anonymous_giving'})
+        response = self.hit_privacy('POST', data={'toggle': 'anonymous_giving'})
+        actual = json.loads(response.body)['anonymous_giving']
+        assert actual is False
+
+    def test_participant_can_toggle_anonymous_receiving(self):
+        response = self.hit_privacy('POST', data={'toggle': 'anonymous_receiving'})
+        actual = json.loads(response.body)
+        assert actual['anonymous_receiving'] is True
+
+    def test_participant_can_toggle_anonymous_receiving_back(self):
+        response = self.hit_privacy('POST', data={'toggle': 'anonymous_receiving'})
+        response = self.hit_privacy('POST', data={'toggle': 'anonymous_receiving'})
+        actual = json.loads(response.body)['anonymous_receiving']
+        assert actual is False
+
+    # Related to is-searchable
 
     def test_meta_robots_tag_added_on_opt_out(self):
         self.hit_privacy('POST', data={'toggle': 'is_searchable'})
@@ -54,3 +81,17 @@ class Tests(Harness):
         alice = Participant.from_username('alice')
         self.make_participant('A-Team', number='plural', is_searchable=False).add_member(alice)
         assert 'A-Team' not in self.client.GET("/explore/teams/").body
+
+    # Related to anonymous-receiving
+
+    def test_team_cannot_toggle_anonymous_receiving(self):
+        self.make_participant('team', claimed_time='now', number='plural')
+        response = self.client.PxST(
+            '/team/privacy.json',
+            auth_as='team',
+            data={'toggle': 'anonymous_receiving'}
+        )
+        actual = response.code
+        expected = 403
+        assert actual == expected
+
